JBRULES-1202 primtive types all return new Long(...) for getValue -now return correct object wrappers.
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@15192 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ kogito-drools/drools-core/src/main/java/org/drools/base/extractors/BaseByteClassFieldExtractor.java
@@ -55,7 +55,7 @@ public abstract class BaseByteClassFieldExtractor extends BaseClassFieldExtracto
     }
 
     public Object getValue(InternalWorkingMemory workingMemory, final Object object) {
-        return new Long( getByteValue( workingMemory, object ) );
+        return new Byte( getByteValue( workingMemory, object ) );
     }
 
     public boolean getBooleanValue(InternalWorkingMemory workingMemory, final Object object) {
